#10 ステータスと優先度のセレクトボックスコンポーネント化
